Put AppEnvironment as separate enum outside rn-config
