+++ template/src/types/react-native-config.d.ts
@@ -1,19 +1,16 @@
-declare module "react-native-config" {
-  export enum AppEnvironment {
-    development = "development",
-    test = "test",
-    staging = "staging",
-    production = "production",
-  }
+import {AppEnvironment} from "./index";
 
+declare module "react-native-config" {
   interface EnvironmentVariables {
     API_URL: string;
     ENVIRONMENT: AppEnvironment;
     IGNORE_NET_INFO: boolean;
   }
 
+  // @ts-ignore
   export const Config: EnvironmentVariables;
 
+  // @ts-ignore
   // eslint-disable-next-line import/no-default-export
   export default Config;
 }
